Change models logic. Fix deleting, favorites. Need to finish top_10 for empty index

+++ movie_finder/finder/views.py
@@ -4,20 +4,22 @@ import requests
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
+from django.db.models import Prefetch
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
 
+from sql_log import query_debugger
 from .forms import CustomUserCreationForm
-from .models import FavoriteMovieUser
+from .models import UserFavorite, Movie
 
 
+@query_debugger
 @login_required(login_url=settings.LOGIN_PAGE_URL)
 def favorites(request):
     if request.method == 'GET':
-        favorite_movies = FavoriteMovieUser.objects.filter(
-            user=request.user).values().order_by('-updated_at')
         return render(request, 'finder/favorites.html', context={
-            'favorite_movies': favorite_movies})
+            'favorite_movies': UserFavorite.get_favorite_movies(
+                request.user)})
     elif request.method == 'POST':
         movie = request.POST.get('movie')
         if movie:
@@ -27,19 +29,23 @@ def favorites(request):
             return HttpResponse("Bad movie data", status=204)
 
 
-def remove_from_favorites_by_imdb_id(request, imdb_id):
-    FavoriteMovieUser.objects.filter(user=request.user,
-                                     imdb_id=imdb_id).delete()
-    return HttpResponse("Removed successful", status=200)
+def favorites_by_imdb_id(request, imdb_id):
+    if request.method == 'DELETE':
+        UserFavorite.objects.filter(user=request.user,
+                                    movie__imdb_id=imdb_id).delete()
+        return HttpResponse("Removed successful", status=200)
 
 
 def add_to_favorites(movie, user):
-    FavoriteMovieUser.objects.create(title=movie['Title'],
-                                     imdb_id=movie['imdbID'],
-                                     year=movie['Year'],
-                                     type=movie['Type'],
-                                     poster=movie['Poster'],
-                                     user=user)
+    try:
+        movie_to_favorite = Movie.objects.get(imdb_id=movie['imdbID'])
+    except Movie.DoesNotExist:
+        movie_to_favorite = Movie.objects.create(title=movie['Title'],
+                                                 imdb_id=movie['imdbID'],
+                                                 year=movie['Year'],
+                                                 type=movie['Type'],
+                                                 poster=movie['Poster'])
+    UserFavorite.objects.create(user=user, movie=movie_to_favorite)
 
 
 def find_movie(movie_to_find):
@@ -61,12 +67,14 @@ def find_movie(movie_to_find):
 
 @login_required(login_url=settings.LOGIN_PAGE_URL)
 def index(request):
-    if request.method == 'POST':
-        movie_to_find = request.POST.get('movie_name_to_find')
-        if movie_to_find:
-            context = find_movie(movie_to_find)
-            return render(request, 'finder/index.html', context=context)
-    return render(request, 'finder/index.html')
+    movie_to_find = request.POST.get('movie_name_to_find')
+    if movie_to_find:
+        context = find_movie(movie_to_find)
+        return render(request, 'finder/index.html', context=context)
+    else:
+        a = Movie.get_top_10()
+        return render(request, 'finder/index.html',
+                      context=Movie.get_top_10())
 
 
 def registration(request):
